src/requirements.txt
#PYTHON VERSION == 3.11.2

absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
Automat==22.10.0
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
constantly==23.10.4
construct==2.10.70
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
dill==0.3.8
executing==2.0.1
ffmpeg-python==0.2.0
filelock==3.13.1
flatbuffers==20181003210633
fonttools==4.48.1
fsspec==2024.2.0
future==0.18.3
gast==0.5.4
google-auth==2.27.0
google-auth-oauthlib==1.2.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
hyperlink==21.0.0
idna==3.6
incremental==22.10.0
install==1.3.5
ipython==8.21.0
jedi==0.19.1
Jinja2==3.1.3
keras==2.15.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.8.2
matplotlib-inline==0.1.6
ml-dtypes==0.2.0
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.9.0.80
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.0
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
pip==23.0.1
ply==3.11
prompt-toolkit==3.0.43
protobuf==4.23.4
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycryptodomex==3.20.0
Pygments==2.17.2
pyparsing==3.1.1
pysmi==0.3.4
pysnmp==4.4.12
python-dateutil==2.8.2
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scipy==1.12.0
seaborn==0.13.2
setuptools==66.1.1
six==1.16.0
stack-data==0.6.3
sympy==1.12
tensor==0.3.6
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tensorflow==2.15.0
tensorflow-cpu-aws==2.15.0
tensorflow-estimator==2.15.0
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tflite-runtime==2.14.0
thop==0.1.1.post2209072238
torch==2.2.0
torchaudio==2.2.0
torchvision==0.17.0
tqdm==4.66.1
traitlets==5.14.1
Twisted==23.10.0
typing_extensions==4.9.0
tzdata==2023.4
ultralytics==8.1.10
urllib3==2.2.0
wcwidth==0.2.13
Werkzeug==3.0.1
wheel==0.42.0
wrapt==1.14.1
zope.interface==6.1
